model/requirements.txt
absl-py==1.0.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.0
alembic==1.11.1
amqp==5.1.1
aniso8601==9.0.1
anyio==3.7.0
apache-beam==2.40.0
apispec==6.0.2
apptools==5.2.1
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==21.4.0
avro-python3==1.10.2
Babel==2.14.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
billiard==4.1.0
bleach==5.0.1
Brotli==1.0.9
cachetools==5.3.0
celery==5.3.4
certifi==2024.2.2
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.1.0
colorama==0.4.6
colour==0.1.5
commonmark==0.9.1
configobj==5.0.8
contextlib2==21.6.0
coremltools==6.3.0
coverage==7.4.1
crcmod==1.7
cryptography==41.0.1
cycler==0.11.0
Cython==0.29.30
databricks-cli==0.17.7
debugpy==1.6.2
decorator==4.4.2
defusedxml==0.7.1
dicomweb-client==0.59.1
diffimg==0.2.3
dill==0.3.1.1
diskcache==5.6.1
distlib==0.3.4
Django==5.0.2
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-money==3.4.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.7
dnspython==1.16.0
docker==6.1.3
docopt==0.6.2
easyocr==1.6.2
ecdsa==0.18.0
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints==0.4
envisage==7.0.3
etils==0.6.0
eventlet @ https://github.com/eventlet/eventlet/archive/master.zip
exceptiongroup==1.2.0
executing==0.8.3
expiring-dict==1.1.0
expiringdict==1.2.2
fastapi==0.95.0
fastavro==1.5.2
fastjsonschema==2.15.3
filelock==3.11.0
fire==0.5.0
Flask==2.1.3
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==4.3.2
flatbuffers==1.12
fonttools==4.31.2
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
gast==0.4.0
gdown==4.7.1
gevent==21.12.0
gevent-websocket==0.10.1
gin-config==0.5.0
girder-client==3.1.17
gitdb==4.0.10
GitPython==3.1.31
glcontext==2.3.5
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
grpcio==1.44.0
h11==0.14.0
h5py==3.6.0
hdfs==2.7.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
huggingface-hub==0.11.1
identify==2.5.24
idna==3.6
imageio==2.19.3
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.8.0
iniconfig==2.0.0
ipykernel==6.15.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isosurfaces==0.1.0
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.12
keras==2.9.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.1
kombu==5.3.2
kraken==4.3.13
labelImg==1.8.6
libclang==13.0.0
lightning-utilities==0.10.0
lmdb==1.4.1
lvis==0.5.3
lxml==4.9.1
Mako==1.2.4
-e git+https://github.com/3b1b/manim.git@bda7f98d2e8b80b053f24d3e6011693b1e89ce76#egg=manimgl
ManimPango==0.4.1
mapbox-earcut==1.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mayavi==4.8.1
mistune==0.8.4
mlflow==2.4.1
moderngl==5.6.4
moderngl-window==2.4.1
monai==1.2.0
monailabel==0.7.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.4
multipledispatch==0.6.0
munch==4.0.0
mutagen==1.46.0
mypy==1.8.0
mypy-extensions==1.0.0
natsort==8.1.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nibabel==5.1.0
ninja==1.11.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.4.12
nptyping==2.5.0
numpy==1.23.5
numpy-stl==3.1.1
numpymaxflow==0.0.5
oauth2client==4.1.3
oauthlib==3.2.0
object-detection @ file:///C:/Additional%20Packages/TFOD/models/research
opencv-python==4.8.0.76
openfabric-pysdk==0.1.11
openslide-python==1.1.2
opt-einsum==3.3.0
orjson==3.7.7
packaging==23.2
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleDB==0.9.2
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==2.5.1
pluggy==1.4.0
poetry-core==1.5.0
pooch==1.8.0
portalocker==2.5.1
pprintpp==0.4.0
pre-commit==3.3.3
pretrainedmodels==0.7.4
proglog==0.1.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.36
proto-plus==1.20.6
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.9.9
pure-eval==0.2.2
py-cpuinfo==8.0.0
py-moneyed==3.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyclipper==1.3.0.post4
pycocotools==2.0.4
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
pydicom==2.3.1
pydicom-seg==0.4.1
PyDirectInput==1.0.4
pydot==1.4.2
pydub==0.25.1
pyface==8.0.0
PyGetWindow==0.0.9
pyglet==1.5.23
Pygments==2.11.2
PyJWT==2.8.0
pymeshfix==0.16.2
pymongo==3.12.3
PyMsgBox==1.0.9
pynetdicom==2.0.2
pynput==1.7.6
pynrrd==1.0.0
PyOpenGL==3.1.6
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.2.0
pyrr==0.10.3
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-engineio==3.14.2
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==6.1.2
python-socketio==4.6.1
python-utils==3.8.1
pytorch-ignite==0.4.11
pytorch-lightning==2.0.9.post0
pytube==15.0.0
pytweening==1.0.4
pytz==2024.1
pyvista==0.42.3
PyWavelets==1.3.0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
qudida==0.0.4
querystring-parser==1.2.4
questionary==2.0.0
rake-nltk==1.0.6
redis==5.0.1
regex==2022.7.9
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retrying==1.3.4
rfc3986==1.5.0
rich==12.2.0
rsa==4.8
runstats==2.0.0
sacrebleu==2.1.0
safetensors==0.3.3
schedule==1.1.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
scooby==0.9.2
screeninfo==0.8
seaborn==0.11.2
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
Shapely==1.8.5.post1
SimpleITK==2.2.1
six==1.16.0
skia-pathops==0.7.2
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sparse-cubes==0.1.0
SpeechRecognition==3.9.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
stack-data==0.2.0
starlette==0.26.1
svgelements==1.6.13
sympy==1.10.1
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-addons==0.17.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.9.0
termcolor==1.1.0
terminado==0.15.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.9.2
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
timeloop==1.0.2
timm==0.9.2
tinycss2==1.1.1
tldextract==5.1.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchaudio==0.13.1
torchmetrics==1.3.0
torchvision==0.15.2
tornado==6.2
tqdm==4.63.1
traitlets==5.3.0
traits==6.4.3
traitsui==8.0.0
transformers==4.25.1
trdg==1.8.0
trimesh==4.0.5
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
uuid==1.30
uvicorn==0.21.1
validators==0.18.2
vine==5.0.0
virtualenv==20.13.2
vtk==9.3.0
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.5
webargs==8.2.0
webencodings==0.5.1
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wikipedia==1.4.0
wrapt==1.14.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2023.7.6
zipp==3.8.1
zope.event==4.6
zope.interface==5.5.2
